[ci] Wrap listUnassignedRewardCouponsV2 with waitUntilAcsIngested in each DB stor
