+++ apps/wallet/src/main/scala/org/lfdecentralizedtrust/splice/wallet/store/db/DbExternalPartyWalletStore.scala
@@ -3,6 +3,7 @@
 
 package org.lfdecentralizedtrust.splice.wallet.store.db
 
+import org.lfdecentralizedtrust.splice.codegen.java.splice.{amulet as amuletCodegen}
 import org.lfdecentralizedtrust.splice.codegen.java.splice.validatorlicense as validatorCodegen
 import org.lfdecentralizedtrust.splice.codegen.java.splice.round.IssuingMiningRound
 import org.lfdecentralizedtrust.splice.codegen.java.splice.types.Round
@@ -16,11 +17,12 @@ import org.lfdecentralizedtrust.splice.store.db.{
   DbTransferInputQueries,
 }
 import org.lfdecentralizedtrust.splice.store.{Limit, LimitHelpers}
-import org.lfdecentralizedtrust.splice.util.{Contract, TemplateJsonDecoder}
+import org.lfdecentralizedtrust.splice.util.{Contract, ContractWithState, TemplateJsonDecoder}
 import org.lfdecentralizedtrust.splice.wallet.store.ExternalPartyWalletStore
 import com.digitalasset.canton.lifecycle.CloseContext
 import com.digitalasset.canton.logging.NamedLoggerFactory
 import com.digitalasset.canton.resource.DbStorage
+import slick.jdbc.canton.ActionBasedSQLInterpolation.Implicits.actionBasedSQLInterpolationCanton
 import com.digitalasset.canton.tracing.TraceContext
 import com.digitalasset.canton.util.ShowUtil.*
 import com.digitalasset.canton.topology.ParticipantId
@@ -75,6 +77,8 @@ class DbExternalPartyWalletStore(
     with LimitHelpers {
 
   import org.lfdecentralizedtrust.splice.store.db.AcsQueries.AcsStoreId
+  import org.lfdecentralizedtrust.splice.util.FutureUnlessShutdownUtil.futureUnlessShutdownToFuture
+  import multiDomainAcsStore.waitUntilAcsIngested
 
   override protected def acsStoreId: AcsStoreId = multiDomainAcsStore.acsStoreId
   override protected def acsTableName: String = WalletTables.externalPartyAcsTableName
@@ -87,6 +91,27 @@ class DbExternalPartyWalletStore(
     org.lfdecentralizedtrust.splice.wallet.store.db.WalletTables.ExternalPartyWalletAcsStoreRowData,
     AcsInterfaceViewRowData.NoInterfacesIngested,
   ] = ExternalPartyWalletStore.contractFilter(key)
+
+  override def listUnassignedRewardCouponsV2(
+      limit: Limit = defaultLimit
+  )(implicit tc: TraceContext): Future[Seq[
+    ContractWithState[amuletCodegen.RewardCouponV2.ContractId, amuletCodegen.RewardCouponV2]
+  ]] =
+    waitUntilAcsIngested {
+      for {
+        result <- storage.query(
+          selectFromAcsTableWithState(
+            WalletTables.externalPartyAcsTableName,
+            acsStoreId,
+            domainMigrationId,
+            amuletCodegen.RewardCouponV2.COMPANION,
+            additionalWhere = sql"and acs.create_arguments->>'beneficiary' is null",
+            orderLimit = sql"order by acs.contract_expires_at asc limit ${sqlLimit(limit)}",
+          ),
+          "listUnassignedRewardCouponsV2",
+        )
+      } yield result.map(contractWithStateFromRow(amuletCodegen.RewardCouponV2.COMPANION)(_))
+    }
 
   override def listSortedLivenessActivityRecords(
       issuingRoundsMap: Map[Round, IssuingMiningRound],
